Major overhaul of table support for improved sorting and copy/paste.
git-svn-id: https://svn.code.sf.net/p/controlremote/code/trunk@56 bfc585dc-5255-483b-a5ca-a9bf5897b33d

+++ km/com/hifiremote/jp1/PanasonicVcr.java
@@ -9,8 +9,8 @@ public class PanasonicVcr
 
   public void initialize( DeviceParameter[] devParms, CmdParameter[] cmdParms )
   {
-    int dev = devParms[0].getValueOrDefault().intValue();
-    int sub = devParms[1].getValueOrDefault().intValue();
+    int dev = (( Integer )devParms[0].getValueOrDefault()).intValue();
+    int sub = (( Integer )devParms[1].getValueOrDefault()).intValue();
     ChoiceCmdParm devCmdParm = ( ChoiceCmdParm )cmdParms[ 0 ];
     ChoiceCmdParm subCmdParm = ( ChoiceCmdParm )cmdParms[ 1 ];
     Choice[] devChoices = devCmdParm.getChoices();
